Components related to main view have no css attribute jmix-framework/jmix_2517
