+++ jmix-flowui/flowui/src/main/java/io/jmix/flowui/xml/layout/support/ComponentLoaderSupport.java
@@ -61,7 +61,10 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
+import java.util.function.BiConsumer;
 import java.util.function.Consumer;
+
+import static org.apache.commons.lang3.StringUtils.trimToEmpty;
 
 @Component("flowui_ComponentLoaderSupport")
 @Scope(BeanDefinition.SCOPE_PROTOTYPE)
@@ -234,6 +237,10 @@ public class ComponentLoaderSupport implements ApplicationContextAware {
 
     public void loadTabIndex(Focusable<?> component, Element element) {
         loaderSupport.loadInteger(element, "tabIndex", component::setTabIndex);
+    }
+
+    public void loadCss(com.vaadin.flow.component.Component component, Element element) {
+        loaderSupport.loadString(element, "css", css -> applyCss(css, component.getStyle()::set));
     }
 
     public void loadThemeNames(HasTheme component, Element element) {
@@ -626,5 +633,25 @@ public class ComponentLoaderSupport implements ApplicationContextAware {
                 }
             }
         }
+    }
+
+    protected void applyCss(String css, BiConsumer<String, String> setter) {
+        Arrays.stream(StringUtils.split(css, ';'))
+                .filter(StringUtils::isNotBlank)
+                .forEach(propertyStatement -> {
+                    int separatorIndex = propertyStatement.indexOf(':');
+                    if (separatorIndex < 0) {
+                        throw new GuiDevelopmentException("Incorrect CSS string: " + css, context);
+                    }
+
+                    String propertyName = trimToEmpty(propertyStatement.substring(0, separatorIndex));
+                    String propertyValue = trimToEmpty(propertyStatement.substring(separatorIndex + 1));
+
+                    if (StringUtils.isBlank(propertyName)) {
+                        throw new GuiDevelopmentException("Incorrect CSS string, empty property name: " + css, context);
+                    }
+
+                    setter.accept(propertyName, propertyValue);
+                });
     }
 }
