+++ jmix-flowui/flowui/src/main/resources/io/jmix/flowui/view/main/main-view-layout.xsd
@@ -26,9 +26,12 @@
         <xs:group ref="layout:singleLayoutOrComponent"/>
 
         <xs:attribute name="touchOptimized" type="xs:boolean"/>
+        <xs:attribute name="css" type="xs:string"/>
     </xs:complexType>
 
     <xs:complexType name="drawerLayoutContainer">
         <xs:group ref="layout:singleLayoutOrComponent"/>
+
+        <xs:attribute name="css" type="xs:string"/>
     </xs:complexType>
 </xs:schema>
