+++ jmix-flowui/flowui/src/main/resources/io/jmix/flowui/view/main/main-view.xsd
@@ -56,6 +56,7 @@
             <xs:element name="drawerLayout" type="mainLayout:drawerLayoutContainer"/>
         </xs:all>
         <xs:attribute name="drawerOpened" type="xs:boolean"/>
+        <xs:attribute name="css" type="xs:string"/>
         <xs:attribute name="primarySection" type="primarySection"/>
 
         <xs:attributeGroup ref="layout:hasClassNames"/>
